[pam_bt_trust] Split code from header
